requirements.txt
# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
mysql-connector-python>=8.0.33

# Deep Learning for Face Recognition
deepface>=0.0.79
tensorflow>=2.13.0

# YOLO for Face Detection
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# GUI
tk>=0.1.0
